Move the components to component dirctory
