+++ client/src/App.js
@@ -2,24 +2,19 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import useApplicationData from './hooks/useApplicationData';
 import CourseList from './components/CourseList';
 
-import Navbar from './NavBar';
-import Signup from './Signup';
-import Login from './Login';
-import Home from './Home';
+import Navbar from './components/NavBar';
+import Signup from './components/Signup';
+import Login from './components/Login';
+import Home from './components/Home';
 
 
 function App() {
 
   const { state, dispatch } = useApplicationData();
-  // const userList = state.users.map((user) => (<li key={user.id} > {user.first_name} {user.last_name} {user.email} </li>
-  // ));
 
   return (
     <div>
-      {/* <h1> Users </h1>
-      <ul> {userList} </ul> */}
       <Router>
-        <CourseList courses={state.courses} dispatch={dispatch} adding={state.addingCourse} />
         <CourseList courses={state.courses} dispatch={dispatch} adding={state.addingCourse} />
         <Navbar state={state} dispatch={dispatch} />
         <Switch>
